ChronoSageEngine/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
numpy==1.26.2
dowhy==0.11.1
networkx==3.2.1
plotly==5.18.0
dash==2.14.2
dash-bootstrap-components==1.5.0
pyyaml==6.0.1
scikit-learn==1.3.2
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.0
scipy==1.11.4
statsmodels==0.14.0
matplotlib==3.8.2
